LayoutTests/discovery/RUIHPL-7.9.3.1.1.g-expected.txt

Tests DLNA Player for AV Media Format. RUIHPL-7.9.3.1: MPEG_TS_NA_ISO MP2 MP@ML, MP@HL(1080p) AC3

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Video Track is playable: /home/gar/MediaContent/O-MP2TS_HN_I-1.mpg
PASS successfullyParsed is true

TEST COMPLETE
